# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Solar Energy Assistant API
  version: 1.0.0
  description: Local single-owner API for owner onboarding, installations, datasets, training, predictions, dashboard, and recommendations.
servers:
  - url: http://localhost:8000
...
paths:
  /owner:
    get:
      tags: [owner]
      operationId: get_owner
      responses:
        '200':
          description: Current local owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerResult'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [owner]
      operationId: create_owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerPayload'
      responses:
        '200':
          description: Owner created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Local owner already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [owner]
      operationId: update_owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerPayload'
      responses:
        '200':
          description: Owner updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /installations/{installation_id}/predictions:
    get:
      tags: [installations]
      operationId: predict_installation_solar_yield
      parameters:
        - $ref: '#/components/parameters/InstallationId'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Prediction result or not-ready state
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PredictionResult'
                  - $ref: '#/components/schemas/PredictionNotReadyResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /models/train:
    post:
      tags: [models]
      operationId: train_model
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [parquet_path]
              properties:
                parquet_path:
                  type: string
                optimize:
                  type: boolean
                  default: false
                advanced_optimization:
                  type: boolean
                  default: false
                n_trials:
                  type: integer
                  default: 50
                activate_model:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Training queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingRunQueuedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Training already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /models/training-runs/{training_run_id}:
    get:
      tags: [models]
      operationId: get_training_run_status
      parameters:
        - in: path
          name: training_run_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Training run status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingRunStatusResult'
        '404':
          $ref: '#/components/responses/NotFound'
  /models:
    get:
      tags: [models]
      operationId: list_models
      parameters:
        - in: query
          name: installation_id
          schema:
            type: string
        - in: query
          name: is_active
          schema:
            type: boolean
        - in: query
          name: target_column
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelResult'
  /models/{model_id}:
    get:
      tags: [models]
      operationId: get_model
      parameters:
        - $ref: '#/components/parameters/ModelId'
      responses:
        '200':
          description: Model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResult'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /models/{model_id}/activate:
    patch:
      tags: [models]
      operationId: activate_model
      parameters:
        - $ref: '#/components/parameters/ModelId'
      responses:
        '200':
          description: Activated model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResult'
        '404':
          $ref: '#/components/responses/NotFound'
  /predictions:
    get:
      tags: [predictions]
      operationId: predict_solar_yield
      parameters:
        - in: query
          name: installation_id
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Prediction result or not-ready state
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PredictionResult'
                  - $ref: '#/components/schemas/PredictionNotReadyResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    InstallationId:
      in: path
      name: installation_id
      required: true
      schema:
        type: string
    ModelId:
      in: path
      name: model_id
      required: true
      schema:
        type: string
    StartDate:
      in: query
      name: start_date
      required: true
      schema:
        type: string
        format: date
    EndDate:
      in: query
      name: end_date
      required: true
      schema:
        type: string
        format: date
  responses:
    BadRequest:
      description: Validation or request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
    OwnerPayload:
      type: object
      required: [first_name]
      properties:
        first_name:
          type: string
        last_name:
          type: string
          nullable: true
    OwnerResult:
      type: object
      required: [id, first_name, display_name, created_at, updated_at]
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
          nullable: true
        display_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TrainingRunQueuedResult:
      type: object
      required: [training_run_id, status]
      properties:
        training_run_id:
          type: string
        status:
          type: string
    TrainingRunSummaryResult:
      type: object
      required: [records, period_start, period_end, best_window_start, best_window_end, average_daily_production_kwh]
      properties:
        records:
          type: integer
        period_start:
          type: string
        period_end:
          type: string
        best_window_start:
          type: string
        best_window_end:
          type: string
        average_daily_production_kwh:
          type: number
          format: double
    TrainingRunStatusResult:
      type: object
      required: [training_run_id, status, phase, progress]
      properties:
        training_run_id:
          type: string
        status:
          type: string
        phase:
          type: string
        progress:
          type: integer
        current_trial:
          type: integer
          nullable: true
        total_trials:
          type: integer
          nullable: true
        best_rmse:
          type: number
          format: double
          nullable: true
        error_code:
          type: string
          nullable: true
        summary:
          allOf:
            - $ref: '#/components/schemas/TrainingRunSummaryResult'
          nullable: true
    ModelResult:
      type: object
      required:
        [id, dataset_id, installation_id, name, model_name, version, model_path, is_active, created_at, mae, rmse, r2, mse, train_size, validation_size, test_size, target_column, metrics]
      properties:
        id:
          type: string
        dataset_id:
          type: string
        installation_id:
          type: string
        dataset_path:
          type: string
          nullable: true
        name:
          type: string
        model_name:
          type: string
        version:
          type: string
        model_path:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        accuracy:
          type: number
          format: double
          nullable: true
        offset_kw:
          type: number
          format: double
          nullable: true
        mae:
          type: number
          format: double
        rmse:
          type: number
          format: double
        r2:
          type: number
          format: double
        mse:
          type: number
          format: double
        train_size:
          type: integer
        validation_size:
          type: integer
        test_size:
          type: integer
        target_column:
          type: string
        metrics:
          type: object
          additionalProperties:
            type: number
            format: double
...
    PredictionPointResult:
      type: object
      required: [timestamp, value]
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
          format: double
    PredictionResult:
      type: object
      required: [unit, total_average, predictions]
      properties:
        unit:
          type: string
        total_average:
          type: number
          format: double
        predictions:
          type: array
          items:
            $ref: '#/components/schemas/PredictionPointResult'
    PredictionNotReadyResult:
      type: object
      required: [status, reason, training_run_id]
      properties:
        status:
          type: string
        reason:
          type: string
        training_run_id:
          type: string
...